# src/main/gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API REST - PROG3 TD Final
    - **A** : Ouverture d'une nouvelle collectivite
    - **B-2** : Admission d'un nouveau membre (conditions renforcees)
    - **J** : Attribution d'un numero et d'un nom unique a une collectivite
    - **C** : Cotisations et paiements
    - **D** : Situation de tresorerie (comptes)
  title: Federation de Collectivites Agricoles API
  version: 0.0.3
servers:
- description: Serveur local
  url: http://localhost:8080
...
paths:
...
  /collectivities/{id}/identity:
    patch:
      description: |
        Une fois le numero ou le nom attribue, il ne peut plus etre modifie.
        Toute tentative de modification retourne 409 Conflict.
      operationId: assignerIdentite
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignIdentityRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectiviteResponse'
          description: Identite attribuee
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
      summary: J - Attribuer numero et nom unique a une collectivite
      tags:
      - Collectivites
      x-content-type: application/json
      x-accepts:
      - application/json
  /members:
    post:
      description: |
        Nouvelles conditions d'admission :
        1. Au moins 2 parrains confirmes (anciennete > 90 jours)
        2. Parrains de la collectivite cible >= parrains exterieurs
        3. Montant = 50 000 MGA (frais adhesion) + cotisations annuelles de la collectivite
        4. Mode de paiement : VIREMENT_BANCAIRE ou MOBILE_MONEY uniquement
        5. Nature de la relation avec chaque parrain obligatoire
      operationId: admettreMembre
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembreRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembreResponse'
          description: Membre admis (poste MEMBRE_JUNIOR attribue automatiquement)
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflit - ressource deja existante ou valeur immuable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Regle metier non respectee
      summary: B-2 - Admettre un nouveau membre
      tags:
      - Membres
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /collectivities/{collectiviteId}/members:
    get:
      operationId: getMembresByCollectivite
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MembreResponse'
                type: array
          description: Liste des membres
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Lister les membres d'une collectivite
      tags:
      - Membres
      x-accepts:
      - application/json
  /members/{membreId}/paiements:
    get:
      operationId: getPaiementsMembre
      parameters:
      - explode: false
        in: path
        name: membreId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PaiementResponse'
                type: array
          description: Liste des paiements
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Historique des paiements d'un membre
      tags:
      - Paiements
      x-accepts:
      - application/json
  /collectivities/{collectiviteId}/cotisations:
    get:
      operationId: getCotisations
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CotisationResponse'
                type: array
          description: Liste des cotisations
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Lister les cotisations d'une collectivite
      tags:
      - Cotisations
      x-accepts:
      - application/json
    post:
      description: |
        Les cotisations peuvent etre MENSUELLE, ANNUELLE ou PONCTUELLE.
      operationId: creerCotisation
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCotisationRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CotisationResponse'
          description: Cotisation creee
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: C - Creer une cotisation pour une collectivite
      tags:
      - Cotisations
      x-content-type: application/json
      x-accepts:
      - application/json
  /collectivities/{collectiviteId}/cotisations/{cotisationId}:
    get:
      operationId: getCotisation
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: cotisationId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CotisationResponse'
          description: Cotisation trouvee
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
      summary: Recuperer une cotisation par ID
      tags:
      - Cotisations
      x-accepts:
      - application/json
...
  /collectivities/{collectiviteId}/paiements:
    get:
      operationId: getPaiements
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - example: 2026-01-01
        explode: true
        in: query
        name: debut
        required: true
        schema:
          format: date
          type: string
        style: form
      - example: 2026-12-31
        explode: true
        in: query
        name: fin
        required: true
        schema:
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PaiementResponse'
                type: array
          description: Liste des paiements
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Regle metier non respectee
      summary: C - Paiements d'une collectivite sur une periode
      tags:
      - Paiements
      x-accepts:
      - application/json
    post:
      description: |
        Le tresorier enregistre un paiement d'un membre pour une cotisation.
        Trace : montant, date d'encaissement, mode de paiement (ESPECE, VIREMENT_BANCAIRE, MOBILE_MONEY).
      operationId: enregistrerPaiement
      parameters:
      - explode: false
        in: path
        name: collectiviteId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaiementRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaiementResponse'
          description: Paiement enregistre
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
          description: Donnees invalides
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource introuvable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Regle metier non respectee
      summary: C - Enregistrer un encaissement (le tresorier)
      tags:
      - Paiements
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
...
  schemas:
...
    AssignIdentityRequest:
      example:
        numero: COL-001
        nom: Collectivite Verte d'Antananarivo
      properties:
        numero:
          description: Numero unique attribue par la federation (immuable)
          example: COL-001
          type: string
        nom:
          description: Nom unique attribue par la federation (immuable)
          example: Collectivite Verte d'Antananarivo
          type: string
      required:
      - nom
      - numero
      type: object
    CreateMembreRequest:
      example:
        parrains:
        - parrainId: 6
          relation: famille
        - parrainId: 6
          relation: famille
        dateNaissance: 2000-01-23
        genre: MASCULIN
        adresse: adresse
        modePaiement: VIREMENT_BANCAIRE
        telephone: telephone
        collectiviteId: 0
        nom: nom
        prenom: prenom
        email: email
        montantPaye: 250000
        metier: metier
      properties:
        collectiviteId:
          format: int64
          type: integer
        nom:
          type: string
        prenom:
          type: string
        dateNaissance:
          format: date
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        adresse:
          type: string
        metier:
          type: string
        telephone:
          type: string
        email:
          format: email
          type: string
        parrains:
          items:
            $ref: '#/components/schemas/ParrainRequest'
          minItems: 2
          type: array
        montantPaye:
          description: 50 000 MGA frais adhesion + cotisations annuelles de la collectivite
          example: 250000
          format: int64
          type: integer
        modePaiement:
          enum:
          - VIREMENT_BANCAIRE
          - MOBILE_MONEY
          type: string
      required:
      - adresse
      - collectiviteId
      - dateNaissance
      - email
      - genre
      - metier
      - modePaiement
      - montantPaye
      - nom
      - parrains
      - prenom
      - telephone
      type: object
    ParrainRequest:
      example:
        parrainId: 6
        relation: famille
      properties:
        parrainId:
          format: int64
          type: integer
        relation:
          description: "Nature de la relation (famille, ami, collegue, etc.)"
          example: famille
          type: string
      required:
      - parrainId
      - relation
      type: object
    CreateCotisationRequest:
      example:
        description: description
        montant: 20000
        typeCotisation: MENSUELLE
      properties:
        typeCotisation:
          $ref: '#/components/schemas/TypeCotisation'
        montant:
          example: 20000
          format: int64
          type: integer
        description:
          type: string
      required:
      - montant
      - typeCotisation
      type: object
    CreatePaiementRequest:
      example:
        cotisationId: 6
        modePaiement: ESPECE
        montant: 1
        membreId: 0
        dateEncaissement: 2000-01-23
      properties:
        membreId:
          format: int64
          type: integer
        cotisationId:
          format: int64
          type: integer
        montant:
          format: int64
          type: integer
        dateEncaissement:
          format: date
          type: string
        modePaiement:
          $ref: '#/components/schemas/ModePaiement'
      required:
      - cotisationId
      - dateEncaissement
      - membreId
      - modePaiement
      - montant
      type: object
...
    CollectiviteResponse:
      example:
        cotisationAnnuelleObligatoire: 6
        ville: ville
        dateCreation: 2000-01-23
        specialiteAgricole: specialiteAgricole
        numero: numero
        id: 0
        nombreMembresActifs: 1
        nom: nom
      properties:
        id:
          format: int64
          type: integer
        numero:
          description: Null jusqu'a l'attribution via PATCH /identity
          nullable: true
          type: string
        nom:
          nullable: true
          type: string
        ville:
          type: string
        specialiteAgricole:
          type: string
        dateCreation:
          format: date
          type: string
        cotisationAnnuelleObligatoire:
          format: int64
          type: integer
        nombreMembresActifs:
          format: int64
          type: integer
      type: object
    MembreResponse:
      example:
        dateNaissance: 2000-01-23
        telephone: telephone
        actif: true
        nom: nom
        dateAdhesion: 2000-01-23
        collectiviteNom: collectiviteNom
        genre: MASCULIN
        adresse: adresse
        id: 0
        poste: PRESIDENT
        collectiviteId: 6
        prenom: prenom
        email: email
        metier: metier
      properties:
        id:
          format: int64
          type: integer
        nom:
          type: string
        prenom:
          type: string
        dateNaissance:
          format: date
          type: string
        genre:
          $ref: '#/components/schemas/Genre'
        adresse:
          type: string
        metier:
          type: string
        telephone:
          type: string
        email:
          type: string
        dateAdhesion:
          format: date
          type: string
        poste:
          $ref: '#/components/schemas/Poste'
        actif:
          type: boolean
        collectiviteId:
          format: int64
          type: integer
        collectiviteNom:
          nullable: true
          type: string
      type: object
    CotisationResponse:
      example:
        collectiviteNom: collectiviteNom
        description: description
        active: true
        montant: 6
        id: 0
        typeCotisation: MENSUELLE
        collectiviteId: 1
      properties:
        id:
          format: int64
          type: integer
        typeCotisation:
          $ref: '#/components/schemas/TypeCotisation'
        montant:
          format: int64
          type: integer
        description:
          type: string
        active:
          type: boolean
        collectiviteId:
          format: int64
          type: integer
        collectiviteNom:
          nullable: true
          type: string
      type: object
    PaiementResponse:
      example:
        cotisationDescription: cotisationDescription
        membreNomPrenom: membreNomPrenom
        modePaiement: ESPECE
        cotisationId: 5
        montant: 6
        id: 0
        membreId: 1
        dateEncaissement: 2000-01-23
      properties:
        id:
          format: int64
          type: integer
        montant:
          format: int64
          type: integer
        dateEncaissement:
          format: date
          type: string
        modePaiement:
          $ref: '#/components/schemas/ModePaiement'
        membreId:
          format: int64
          type: integer
        membreNomPrenom:
          type: string
        cotisationId:
          format: int64
          type: integer
        cotisationDescription:
          nullable: true
          type: string
      type: object
...
    ErrorResponse:
      example:
        message: message
        timestamp: 2000-01-23T04:56:07.000+00:00
        status: 0
      properties:
        timestamp:
          format: date-time
          type: string
        status:
          type: integer
        message:
          type: string
      type: object
    ValidationErrorResponse:
      example:
        errors:
        - errors
        - errors
        timestamp: 2000-01-23T04:56:07.000+00:00
        status: 400
      properties:
        timestamp:
          format: date-time
          type: string
        status:
          example: 400
          type: integer
        errors:
          items:
            type: string
          type: array
      type: object
    Genre:
      enum:
      - MASCULIN
      - FEMININ
      type: string
    Poste:
      enum:
      - PRESIDENT
      - PRESIDENT_ADJOINT
      - TRESORIER
      - SECRETAIRE
      - MEMBRE_CONFIRME
      - MEMBRE_JUNIOR
      type: string
    TypeCotisation:
      enum:
      - MENSUELLE
      - ANNUELLE
      - PONCTUELLE
      type: string
    ModePaiement:
      enum:
      - ESPECE
      - VIREMENT_BANCAIRE
      - MOBILE_MONEY
      type: string
...